+++ R/AllGenerics.R
@@ -2,9 +2,10 @@
 #' @include AllClasses.R
 NULL
 
-# S4 dispatch to S3 generics ===================================================
-setGeneric("jackknife", package = "arkhe")
-setGeneric("bootstrap", package = "arkhe")
+# Import S4 generics ===========================================================
+#' @importMethodsFrom arkhe jackknife
+#' @importMethodsFrom arkhe bootstrap
+NULL
 
 # Mutators =====================================================================
 ## Subset ----------------------------------------------------------------------
@@ -861,6 +862,7 @@ setGeneric(
 #'  refinement will be applied over: `1` indicates rows, `2` indicates columns.
 #' @param axes An [`integer`] vector giving the subscripts of the CA axes to be
 #'  used.
+#' @param x A [`RefinePermutationOrder-class`] object
 #' @param ... Further arguments to be passed to internal methods.
 #' @details
 #'  `seriate_refine()` allows to identify samples that are subject to
@@ -877,7 +879,9 @@ setGeneric(
 #'  the CA are highly stable and which produces an ordering consistent with the
 #'  assumptions of frequency seriation."
 #' @return
-#'  A [`RefinePermutationOrder-class`] object.
+#'  * `seriate_refine()` returns a [`RefinePermutationOrder-class`] object.
+#'  * `hist()` is called it for its side-effects: it results in a histogram
+#'    being displayed (invisibly returns `x`).
 #' @references
 #'  Peeples, M. A., & Schachner, G. (2012). Refining correspondence
 #'  analysis-based ceramic seriation of regional data sets. *Journal of